+++ test/test.ads.js
@@ -278,7 +278,6 @@ QUnit.test('"contentupdate" should fire when src is changed in "content-resuming
   this.player.trigger('adtimeout');
   this.player.trigger('ended');
   this.player.trigger('adtimeout');
-  this.player.ads.snapshot.ended = true;
 
   // set src and trigger synthetic 'loadstart'
   this.player.src('http://media.w3.org/2010/05/sintel/trailer.mp4');
@@ -298,7 +297,6 @@ QUnit.test('"contentupdate" should fire when src is changed in "content-playback
   this.player.trigger('adtimeout');
   this.player.trigger('ended');
   this.player.trigger('adtimeout');
-  this.player.ads.snapshot.ended = true;
   this.player.trigger('ended');
 
   // set src and trigger synthetic 'loadstart'
@@ -485,7 +483,6 @@ QUnit.test('adserror in postroll? transitions to content-playback and fires ende
   this.player.trigger('ended');
   assert.strictEqual(this.player.ads.state, 'postroll?');
 
-  this.player.ads.snapshot.ended = true;
   this.player.trigger('adserror');
   assert.strictEqual(this.player.ads.state, 'content-resuming');
   assert.strictEqual(this.player.ads.triggerevent, 'adserror', 'adserror should be the trigger event');
@@ -506,7 +503,6 @@ QUnit.test('adtimeout in postroll? transitions to content-playback and fires end
   this.player.trigger('ended');
   assert.strictEqual(this.player.ads.state, 'postroll?');
 
-  this.player.ads.snapshot.ended = true;
   this.player.trigger('adtimeout');
   assert.strictEqual(this.player.ads.state, 'content-resuming');
   assert.strictEqual(this.player.ads.triggerevent, 'adtimeout', 'adtimeout should be the trigger event');
@@ -527,7 +523,6 @@ QUnit.test('adskip in postroll? transitions to content-playback and fires ended'
   this.player.trigger('ended');
   assert.strictEqual(this.player.ads.state, 'postroll?');
 
-  this.player.ads.snapshot.ended = true;
   this.player.trigger('adskip');
   assert.strictEqual(this.player.ads.state, 'content-resuming');
   assert.strictEqual(this.player.ads.triggerevent, 'adskip', 'adskip should be the trigger event');
@@ -553,7 +548,6 @@ QUnit.test('an "ended" event is fired in "content-resuming" via a timeout if not
   assert.strictEqual(this.player.ads.state, 'postroll?');
 
   this.player.ads.startLinearAdMode();
-  this.player.ads.snapshot.ended = true;
   this.player.ads.endLinearAdMode();
   assert.strictEqual(this.player.ads.state, 'content-resuming');
   assert.strictEqual(endedSpy.callCount, 0, 'we should not have gotten an ended event yet');
@@ -579,7 +573,6 @@ QUnit.test('an "ended" event is not fired in "content-resuming" via a timeout if
   assert.strictEqual(this.player.ads.state, 'postroll?');
 
   this.player.ads.startLinearAdMode();
-  this.player.ads.snapshot.ended = true;
   this.player.ads.endLinearAdMode();
   assert.strictEqual(this.player.ads.state, 'content-resuming');
   assert.strictEqual(endedSpy.callCount, 0, 'we should not have gotten an ended event yet');
@@ -1169,3 +1162,40 @@ QUnit.test('adserror ends linear ad mode ', function(assert) {
   this.player.trigger('adserror');
   assert.strictEqual(this.player.ads._inLinearAdMode, false, 'after adserror');
 });
+
+if (videojs.browser.IS_IOS) {
+  QUnit.test('Check the textTrackChangeHandler takes effect on iOS', function(assert) {
+    const tracks = this.player.textTracks();
+
+    // Since addTextTrack is async, wait for the addtrack event
+    tracks.on('addtrack', function() {
+
+      // Confirm the track is added, set the mode to showing
+      assert.equal(tracks.length, 1);
+      tracks[0].mode = 'showing';
+      assert.equal(tracks[0].mode, 'showing',  'Initial state is showing');
+
+      // Start the ad, confirm the track is disabled
+      this.player.ads.startLinearAdMode();
+      assert.equal(tracks[0].mode, 'disabled', 'Snapshot sets tracks to disabled');
+
+      // Force the mode to showing
+      tracks[0].mode = 'showing';
+
+    }.bind(this));
+
+    // The mode should go back to disabled when the change event happens as
+    // during ad playback we do not want the content captions to be visible on iOS
+    tracks.on('change', function() {
+      assert.equal(tracks[0].mode, 'disabled', 'Mode is reset to disabled');
+
+      // End the ad, check the track mode is showing again
+      this.player.ads.endLinearAdMode();
+      assert.equal(tracks[0].mode, 'showing', 'Mode is restored after ad');
+    }.bind(this));
+
+    this.player.trigger('play');
+    this.player.trigger('adsready');
+    this.player.addTextTrack('captions', 'English', 'en');
+  });
+}